SSCS-7639 Correcting IssueNoticeHandler base methods,  and injecting ReferenceDataService into AdjournCasePreviewService
