+++ build.gradle
@@ -261,7 +261,7 @@ dependencies {
     compile group: 'uk.gov.hmcts.reform', name: 'logging-appinsights', version: versions.reformLogging
     compile group: 'uk.gov.hmcts.reform', name: 'doc-assembly-client', version: '1.0.5'
     compile group: 'uk.gov.hmcts.reform', name: 'document-management-client', version: '7.0.0'
-    compile group: 'uk.gov.hmcts.reform', name: 'sscs-common', version: '3.3.20'
+    compile group: 'uk.gov.hmcts.reform', name: 'sscs-common', version: '3.3.21-RC-SSCS-7639'
     compile group: 'uk.gov.hmcts.reform', name: 'sscs-pdf-email-common', version: '1.4.6'
 
     testCompile group: 'io.rest-assured', name: 'rest-assured'
